+++ marketplace/services/vtex/generic_service.py
@@ -51,10 +51,10 @@ class VtexService:
         self.product_manager = ProductFacebookManager()
         self.app_manager = AppVtexManager()
 
-    @property
-    def fb_service(self) -> FacebookService:  # pragma: no cover
+    def fb_service(self, app: App) -> FacebookService:  # pragma: no cover
+        access_token = app.apptype.get_access_token(app)
         if not self._fb_service:
-            self._fb_service = self.fb_service_class(self.fb_client_class())
+            self._fb_service = self.fb_service_class(self.fb_client_class(access_token))
         return self._fb_service
 
     def get_private_service(
@@ -139,6 +139,7 @@ class VtexService:
         file_name = f"csv_vtex_products_{current_time}.csv"
         product_feed = self._create_product_feed(file_name, catalog)
         self._upload_product_feed(
+            catalog,
             product_feed.facebook_feed_id,
             products_csv,
             file_name,
@@ -146,9 +147,8 @@ class VtexService:
         return product_feed
 
     def _create_product_feed(self, name, catalog: Catalog) -> ProductFeed:
-        response = self.fb_service.create_product_feed(
-            catalog.facebook_catalog_id, name
-        )
+        service = self.fb_service(catalog.app)
+        response = service.create_product_feed(catalog.facebook_catalog_id, name)
 
         if "id" not in response:
             raise UnexpectedFacebookApiResponseValidationError()
@@ -162,9 +162,10 @@ class VtexService:
         return product_feed
 
     def _upload_product_feed(
-        self, product_feed_id, csv_file, file_name, update_only=False
+        self, catalog, product_feed_id, csv_file, file_name, update_only=False
     ):
-        response = self.fb_service.upload_product_feed(
+        service = self.fb_service(catalog)
+        response = service.upload_product_feed(
             product_feed_id, csv_file, file_name, "text/csv", update_only
         )
         if "id" not in response: